odd even added- changes, max programme

+++ src/JavaProgrammes/Odd_Even.java
@@ -6,7 +6,7 @@ public class Odd_Even {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Enter the number");
-        long user_inp= sc.nextLong();
+        double user_inp= sc.nextDouble();
         if(user_inp%2==0)
         {
             System.out.println("Even");
